+++ Tests/CSharp/StatementTests/StatementTests.cs
@@ -483,8 +483,8 @@ public partial class TestClass
         numArray = new int[4];
         numArray = null;
         numArray2[1] = 1;
-        Array.Resize(numArray, 6);
-        Array.Resize(numArray2, 6);
+        Array.Resize(ref numArray, 6);
+        Array.Resize(ref numArray2, 6);
         var y = new int[7, 6];
         y[2, 3] = 1;
         var oldY = y;
@@ -494,9 +494,7 @@ public partial class TestClass
                 Array.Copy(oldY, i * oldY.GetLength(1), y, i * y.GetLength(1), Math.Min(oldY.GetLength(1), y.GetLength(1)));
         return numArray2;
     }
-}
-1 target compilation errors:
-CS1620: Argument 1 must be passed with the 'ref' keyword");
+}");
         }
 
         [Fact]