+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 
 ### VB -> C#
-
+* When converting ReDim Preserve to Array.Resize, "ref" is now added
 
 ### C# -> VB